requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pinecone-client>=3.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
tqdm>=4.66.0
